+++ lib/domain/steps/step_03_remove_duplicates.dart
@@ -149,7 +149,8 @@ class RemoveDuplicatesStep extends ProcessingStep {
         '\n[Step 3/8] Finding and removing duplicates... (This might take some time!)',
       );
 
-      final int removedCount = context.mediaCollection.removeDuplicates();
+      final int removedCount = await context.mediaCollection
+          .removeDuplicatesAsync();
 
       stopwatch.stop();
       return StepResult.success(